netter-ats-backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic[email]
psycopg[binary]
psycopg_pool

